extract the tabNavigator from the app.js file and make as a component.
