+++ components/TabNavigator.js
@@ -1,0 +1,38 @@
+import React from 'react';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import { FontAwesome5, Ionicons } from '@expo/vector-icons';
+import ActivitiesScreen from '../screens/ActivitiesScreen';
+import DietScreen from '../screens/DietScreen';
+import SettingsScreen from '../screens/SettingsScreen';
+
+const Tab = createBottomTabNavigator();
+
+const TabNavigator = () => {
+  return (
+    <Tab.Navigator>
+      <Tab.Screen
+        name="Activities"
+        component={ActivitiesScreen}
+        options={{
+          tabBarIcon: () => <FontAwesome5 name="running" size={24} color="grey" />,
+        }}
+      />
+      <Tab.Screen
+        name="Diet"
+        component={DietScreen}
+        options={{
+          tabBarIcon: () => <Ionicons name="fast-food" size={24} color="grey" />,
+        }}
+      />
+      <Tab.Screen
+        name="Settings"
+        component={SettingsScreen}
+        options={{
+          tabBarIcon: () => <Ionicons name="settings" size={24} color="grey" />,
+        }}
+      />
+    </Tab.Navigator>
+  );
+};
+
+export default TabNavigator;
